Move signer to new (temporary) gem.
