+++ gems/aws-crt-auth/lib/aws-crt-auth.rb
@@ -5,8 +5,6 @@ require_relative 'aws-crt-auth/credentials'
 require_relative 'aws-crt-auth/signing_config'
 require_relative 'aws-crt-auth/signing_result'
 require_relative 'aws-crt-auth/signable'
-require_relative 'aws-crt-auth/signer'
-require_relative 'aws-crt-auth/errors'
 
 module Aws
   module Crt
