--- gems/aws-crt-auth/spec/signer_spec.rb
@@ -1,321 +1,0 @@
-# frozen_string_literal: true
-
-require_relative 'spec_helper'
-require 'weakref'
-
-module Aws
-  module Crt
-    module Auth #:nodoc:
-      describe Signer do
-        let(:credentials) do
-          {
-            access_key_id: 'akid',
-            secret_access_key: 'secret'
-          }
-        end
-
-        let(:options) do
-          {
-            service: 'SERVICE',
-            region: 'REGION',
-            credentials_provider: StaticCredentialsProvider.new(credentials)
-          }
-        end
-
-        context '#service' do
-          it 'requires a service' do
-            expect do
-              Signer.new(
-                region: 'us-east-1',
-                access_key_id: 'akid',
-                secret_access_key: 'secret'
-              )
-            end.to raise_error(ArgumentError, /:service/)
-          end
-
-          it 'accepts a string' do
-            expect(Signer.new(options).service).to eq(options[:service])
-          end
-        end
-
-        context '#region' do
-          it 'requires a region' do
-            expect do
-              Signer.new(
-                service: 'ec2',
-                access_key_id: 'akid',
-                secret_access_key: 'secret'
-              )
-            end.to raise_error(ArgumentError, /:region/)
-          end
-
-          it 'accepts a string' do
-            expect(Signer.new(options).region).to eq(options[:region])
-          end
-        end
-
-        describe '#credentials' do
-          let(:options) { { service: 'ec2', region: 'us-east-1' } }
-
-          it 'requires credentials' do
-            expect do
-              Signer.new(options)
-            end.to raise_error(ArgumentError, /missing credentials/)
-          end
-
-          it 'accepts :access_key_id and :secret_access_key' do
-            signer = Signer.new(options.merge(
-                                  access_key_id: 'akid',
-                                  secret_access_key: 'secret'
-                                ))
-            creds = signer.credentials_provider.credentials
-            expect(creds.access_key_id).to eq('akid')
-            expect(creds.secret_access_key).to eq('secret')
-            expect(creds.session_token).to be(nil)
-          end
-
-          it 'accepts credentials with a session token' do
-            signer = Signer.new(options.merge(
-                                  access_key_id: 'akid',
-                                  secret_access_key: 'secret',
-                                  session_token: 'token'
-                                ))
-            creds = signer.credentials_provider.credentials
-            expect(creds.access_key_id).to eq('akid')
-            expect(creds.secret_access_key).to eq('secret')
-            expect(creds.session_token).to eq('token')
-          end
-
-          it 'accepts :credentials' do
-            signer = Signer.new(options.merge(
-                                  credentials: Credentials.new('akid', 'secret', 'token')
-                                ))
-            creds = signer.credentials_provider.credentials
-            expect(creds.access_key_id).to eq('akid')
-            expect(creds.secret_access_key).to eq('secret')
-            expect(creds.session_token).to eq('token')
-          end
-
-          it 'accepts :credentials_provider' do
-            signer = Signer.new(options.merge(
-                                  credentials_provider: StaticCredentialsProvider.new(
-                                    access_key_id: 'akid',
-                                    secret_access_key: 'secret',
-                                    session_token: 'token'
-                                  )
-                                ))
-            creds = signer.credentials_provider.credentials
-            expect(creds.access_key_id).to eq('akid')
-            expect(creds.secret_access_key).to eq('secret')
-            expect(creds.session_token).to eq('token')
-          end
-
-          # NOTE: This is a different behavior from SDK Sigv4 Signer
-          it 'does not accept empty credentials' do
-            expect do
-              Signer.new(options.merge(
-                           access_key_id: '',
-                           secret_access_key: ''
-                         ))
-            end.to raise_error(ArgumentError)
-          end
-        end
-
-        context '#sign_request' do
-          let(:request) do
-            {
-              http_method: 'GET',
-              url: 'http://domain.com'
-            }
-          end
-
-          it 'populates the Host header' do
-            signature = Signer.new(options).sign_request(request)
-            expect(signature.headers['host']).to eq('domain.com')
-          end
-
-          context 'when a Host header is provided' do
-            it 'uses the provided Host header' do
-              signature = Signer.new(options).sign_request(
-                request.merge(headers: { 'host' => 'otherdomain.com' })
-              )
-
-              expect(signature.headers['host']).to eql('otherdomain.com')
-            end
-          end
-
-          context 'when URI schema is known' do
-            it 'omits port in Host when port not provided' do
-              signature = Signer.new(options).sign_request(
-                http_method: 'GET',
-                url: 'https://domain.com'
-              )
-              expect(signature.headers['host']).to eq('domain.com')
-            end
-
-            it 'omits port in Host when default port and uri port are the same' do
-              signature = Signer.new(options).sign_request(
-                http_method: 'GET',
-                url: 'https://domain.com:443'
-              )
-              expect(signature.headers['host']).to eq('domain.com')
-            end
-
-            it 'includes port in Host when default port and uri port are different' do
-              signature = Signer.new(options).sign_request(
-                http_method: 'GET',
-                url: 'https://domain.com:123'
-              )
-              expect(signature.headers['host']).to eq('domain.com:123')
-            end
-          end
-
-          context 'when URI schema is unknown' do
-            it 'omits port in Host when uri port not provided' do
-              signature = Signer.new(options).sign_request(
-                http_method: 'GET',
-                url: 'abcd://domain.com'
-              )
-              expect(signature.headers['host']).to eq('domain.com')
-            end
-
-            it 'includes port in Host when uri port provided' do
-              signature = Signer.new(options).sign_request(
-                http_method: 'GET',
-                url: 'abcd://domain.com:123'
-              )
-              expect(signature.headers['host']).to eq('domain.com:123')
-            end
-          end
-
-          it 'sets the X-Amz-Date header' do
-            now = Time.now
-            allow(Time).to receive(:now).and_return(now)
-            signature = Signer.new(options).sign_request(
-              http_method: 'GET',
-              url: 'https://domain.com:123'
-            )
-            expect(signature.headers['x-amz-date']).to eq(now.utc.strftime('%Y%m%dT%H%M%SZ'))
-          end
-
-          it 'uses the X-Amz-Date header of the request if present' do
-            now = Time.now.utc.strftime('%Y%m%dT%H%M%SZ')
-            signature = Signer.new(options).sign_request(
-              http_method: 'GET',
-              url: 'https://domain.com',
-              headers: {
-                'X-Amz-Date' => now
-              }
-            )
-            expect(signature.headers['x-amz-date']).to eq(now)
-          end
-
-          it "populates the 'X-Amz-Security-Token' header with session token" do
-            credentials[:session_token] = 'token'
-            signature = Signer.new(options).sign_request(
-              http_method: 'GET',
-              url: 'https://domain.com'
-            )
-            expect(signature.headers['x-amz-security-token']).to eq('token')
-          end
-
-          it "omits 'X-Amz-Security-Token' header when session token is nil" do
-            credentials.delete(:session_token)
-            signature = Signer.new(options).sign_request(
-              http_method: 'GET',
-              url: 'https://domain.com'
-            )
-            expect(signature.headers['x-amz-security-token']).to be(nil)
-          end
-
-          it 'adds the X-Amz-Content-Sha256 header by default' do
-            signature = Signer.new(options).sign_request(
-              http_method: 'GET',
-              url: 'https://domain.com',
-              body: 'abc'
-            )
-            expect(signature.headers['x-amz-content-sha256']).to eq(Digest::SHA256.hexdigest('abc'))
-          end
-
-          it 'can omit the X-Amz-Content-Sha256 header' do
-            options[:apply_checksum_header] = false
-            signature = Signer.new(options).sign_request(
-              http_method: 'GET',
-              url: 'https://domain.com',
-              body: 'abc'
-            )
-            expect(signature.headers['x-amz-content-sha256']).to be(nil)
-          end
-
-          it 'computes the checksum of files without loading them into memory' do
-            body = Tempfile.new('tempfile')
-            body.write('abc')
-            body.flush
-            expect(body).not_to receive(:read)
-            expect(body).not_to receive(:rewind)
-            signature = Signer.new(options).sign_request(
-              http_method: 'POST',
-              url: 'https://domain.com',
-              body: body
-            )
-            expect(signature.headers['x-amz-content-sha256']).to eq(Digest::SHA256.hexdigest('abc'))
-          end
-
-          it 'reads non-file IO objects into  memory to compute checksusm' do
-            signature = Signer.new(options).sign_request(
-              http_method: 'PUT',
-              url: 'http://domain.com',
-              body: StringIO.new('abc')
-            )
-            expect(signature.content_sha256).to eq(Digest::SHA256.hexdigest('abc'))
-          end
-
-          it 'does not read the body if X-Amz-Content-Sha256 if already present' do
-            body = double('http-payload')
-            expect(body).to_not receive(:read)
-            expect(body).to_not receive(:rewind)
-            signature = Signer.new(options).sign_request(
-              http_method: 'PUT',
-              url: 'http://domain.com',
-              headers: {
-                'X-Amz-Content-Sha256' => 'hexdigest'
-              },
-              body: body
-            )
-            expect(signature.headers['x-amz-content-sha256']).to eq('hexdigest')
-          end
-
-          it "populates the 'Authorization' header" do
-            headers = {}
-            signature = Signer.new(options).sign_request(
-              http_method: 'PUT',
-              url: 'http://domain.com',
-              headers: headers
-            )
-            # applied to the signature headers, not the request
-            expect(headers['authorization']).to be(nil)
-            expect(signature.headers['authorization']).to_not be(nil)
-          end
-
-          it 'signs the request' do
-            options[:unsigned_headers] = ['content-length']
-            signature = Signer.new(options).sign_request(
-              http_method: 'PUT',
-              url: 'https://domain.com',
-              headers: {
-                'Foo' => 'foo',
-                'Bar' => 'bar  bar',
-                # NOTE: Ruby Sigv4 incorrectly keeps spaces in quoted headers
-                'Bar2' => '"bar bar"',
-                'Content-Length' => 9,
-                'X-Amz-Date' => '20120101T112233Z'
-              },
-              body: StringIO.new('http-body')
-            )
-            expect(signature.headers['authorization']).to eq('AWS4-HMAC-SHA256 Credential=akid/20120101/REGION/SERVICE/aws4_request, SignedHeaders=bar;bar2;foo;host;x-amz-content-sha256;x-amz-date, Signature=4bae5054b2e035212a0eb42339a957809a8c9428e628fd4b92e5a295d0fa6e5b')
-          end
-        end
-      end
-    end
-  end
-end
