+++ gems/aws-crt-signer/CHANGELOG.md
@@ -1,0 +1,8 @@
+Unreleased Changes
+------------------
+
+
+0.1.0.pre (2020-10-14)
+------------------
+
+* Feature - Initial preview release of `aws-crt-auth` gem.
