+++ gems/aws-crt-signer/VERSION
@@ -1,0 +1,1 @@
+0.1.0.pre
