+++ gems/aws-crt-signer/lib/aws-crt-signer.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require 'aws-crt-auth'
+require_relative 'aws-crt-signer/signer'
